Defined functions are dropped from 'functions'
Function calls are registered in 'functions' only if those names
are not def'ed in the analyzed piece of code. This is done
as a workaround for
https://github.com/NUS-ALSET/achievements/issues/328. However, since
no static code analysis is actually performed, this can lead to both
false positives and false negatives.

An example of a false positive:

    def foo():
        pass

    bar = foo
    bar()

The function call 'bar()' on the last line actually refers to the
function object defined in the same module, however 'bar' will appear
in the 'functions' features.

An example of a false negative:

    def join():
        pass

    s = "".join(some_list)

The function call '.join()' has no relation to the function defined
above it, however it will be removed from the 'functions' features.